feat: Support for OCI Image Index and Docker Manifest List
OCI Image Index and Docker Manifest List are lists of manifests using
the same tag, but built for different platforms (OS/architecture). This
commit adds support for both types of lists/indexes and a way of
selecting a preferred platform with a fallback to only architecture
similar to how Docker solves this problem.